requirements.txt
streamlit==1.44.1
google-generativeai==0.8.5
psycopg2-binary==2.9.10